+++ janitor-api/src/main/java/com/eischet/janitor/api/Janitor.java
@@ -8,6 +8,7 @@ import com.eischet.janitor.api.types.BuiltinTypes;
 import com.eischet.janitor.api.types.JanitorObject;
 import com.eischet.janitor.api.types.builtin.*;
 import com.eischet.janitor.api.types.dispatch.DispatchTable;
+import com.eischet.janitor.toolbox.json.api.JsonException;
 import org.intellij.lang.annotations.Language;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -23,7 +24,7 @@ import java.util.stream.Stream;
 
 /**
  * This is the main entry point for working with the Janitor language in Java.
- *
+ * <p>
  * You'll use this class to:
  * <ul>
  * <li>Access the basic built-in types NULL, TRUE and FALSE.</li>
@@ -72,21 +73,9 @@ public final class Janitor {
     private static @Nullable JanitorEnvironmentProvider userProvider = null;
 
     /**
-     * Compiles a script (source code) to a runnable script (an executable AST in the default implementation).
-     *
-     * @param runtime a runtime to use [TODO: this is actually only used to retrieve an environment, so this could and should be dropped soon!]
-     * @param moduleName the name of the script module; this will show up in stack traces, so it's a good idea to pick an informative name here if you run lots of scripts
-     * @param source the actual script source code
-     * @return the script in executable form
-     * @throws JanitorCompilerException on compiler errors
+     * Private constructor, to keep you from creating instances of this singleton / "namespace class".
      */
-    public RunnableScript compile(JanitorRuntime runtime, String moduleName, @Language("Janitor") String source) throws JanitorCompilerException {
-        return runtime.compile(moduleName, source);
-    }
-
-    // TODO: properly document this, and when you're at it make this more useful.
-    public RunnableScript checkCompile(JanitorRuntime runtime, String moduleName, @Language("Janitor") String source) throws JanitorCompilerException {
-        return runtime.checkCompile(moduleName, source);
+    private Janitor() {
     }
 
     /**
@@ -125,6 +114,7 @@ public final class Janitor {
 
     /**
      * Return a new JInt object, or NULL if the value is null.
+     *
      * @param value a Java Long or null
      * @return a JInt from the Java Long, or NULL
      */
@@ -138,6 +128,7 @@ public final class Janitor {
 
     /**
      * Return a new JInt object, or NULL if the value is null.
+     *
      * @param value a Java Integer or null
      * @return a JInt from the Java Long, or NULL
      */
@@ -151,6 +142,7 @@ public final class Janitor {
 
     /**
      * Return a new JMap object.
+     *
      * @return a new JMap object
      */
     @NotNull
@@ -160,6 +152,7 @@ public final class Janitor {
 
     /**
      * Return a new empty JList.
+     *
      * @return an empty JList
      */
     @NotNull
@@ -171,6 +164,7 @@ public final class Janitor {
      * Returns a new empty list, preallocated for the specified number of elements.
      * Even though this is probably a negligible optimization, if you know the required list size before creating the list,
      * you can tell us about it here. If you don't have the size, do not bother and use the plain list() method.
+     *
      * @param initialSize the initial size of the list
      * @return an empty JList
      */
@@ -181,6 +175,7 @@ public final class Janitor {
 
     /**
      * Returns a new JList as a copy of a java.util.List of JanitorObject elements.
+     *
      * @param list a list of JanitorObject elements
      * @return a JList containing all elements of the list argument
      */
@@ -191,6 +186,7 @@ public final class Janitor {
 
     /**
      * Returns a new JList with all elements from the Stream of JanitorObject elements.
+     *
      * @param stream a stram of JanitorObject elements
      * @return a JList containing all elements of the stream argument
      */
@@ -202,7 +198,8 @@ public final class Janitor {
     /**
      * Returns a new JList with all elements from the Stream of JanitorObject elements.
      * Use this when returning a writable list for an object property.
-     * @param stream a stream of JanitorObject elements
+     *
+     * @param stream   a stream of JanitorObject elements
      * @param onUpdate will be called whenever the JList object is changed
      * @return a JList containing all elements of the stream argument
      */
@@ -213,6 +210,7 @@ public final class Janitor {
 
     /**
      * Returns a new empty JSet.
+     *
      * @return a new empty JSet
      */
     @NotNull
@@ -222,6 +220,7 @@ public final class Janitor {
 
     /**
      * Returns a new JSet containing all elements of the provided collection.
+     *
      * @param collection a collection of JanitorObject elements
      * @return a JSet containing all elements of the collection
      */
@@ -231,6 +230,7 @@ public final class Janitor {
 
     /**
      * Returns a new JSet containing all elements of the provided stream.
+     *
      * @param stream a stream of JanitorObject elements
      * @return a JSet containing all elements of the stream
      */
@@ -240,6 +240,7 @@ public final class Janitor {
 
     /**
      * Returns a new JInt from the long argument.
+     *
      * @param value a number
      * @return a JInt representing the number
      */
@@ -249,6 +250,7 @@ public final class Janitor {
 
     /**
      * Returns a new JInt from the int argument.
+     *
      * @param value a number
      * @return a JInt representing the number
      */
@@ -258,6 +260,7 @@ public final class Janitor {
 
     /**
      * Returns a new JInt from the Number, or NULL in case Java null is passed in.
+     *
      * @param value a number or null
      * @return a JInt representing the number or NULL
      */
@@ -267,6 +270,7 @@ public final class Janitor {
 
     /**
      * Returns a new JBinary object from the byte array.
+     *
      * @param arr a byte array
      * @return a JBinary object
      */
@@ -276,6 +280,7 @@ public final class Janitor {
 
     /**
      * Returns a new JFloat from the passed in value, or NULL if Java null is passed.
+     *
      * @param value Java null or a Double
      * @return NULL or a JFloat representing the argument
      */
@@ -285,6 +290,7 @@ public final class Janitor {
 
     /**
      * Returns a new JFloat from the passed in value.
+     *
      * @param value a double
      * @return a JFloat representing the argument
      */
@@ -294,6 +300,7 @@ public final class Janitor {
 
     /**
      * Returns a new JFloat from the passed in value.
+     *
      * @param value a long
      * @return a JFloat representing the argument
      */
@@ -303,6 +310,7 @@ public final class Janitor {
 
     /**
      * Returns a new JFloat from the passed in value.
+     *
      * @param value an int
      * @return a JFloat representing the argument
      */
@@ -310,11 +318,11 @@ public final class Janitor {
         return getBuiltins().floatingPoint(value);
     }
 
-
     /**
      * Returns a new JDuration from the arguments.
+     *
      * @param value the number of e.g. seconds, minutes, hours...
-     * @param kind the unit, e.g. seconds, minutes, hours
+     * @param kind  the unit, e.g. seconds, minutes, hours
      * @return the JDuration
      */
     public static @NotNull JDuration duration(long value, JDuration.JDurationKind kind) {
@@ -323,6 +331,7 @@ public final class Janitor {
 
     /**
      * Returns a new JRegex for the Java regular expression pattern.
+     *
      * @param pattern a Java Pattern
      * @return a JRegex instance
      */
@@ -331,7 +340,18 @@ public final class Janitor {
     }
 
     /**
+     * Returns a valid Janitor object for the nullable parameter, i.e. turns Java null into Janitor.NULL.
+     *
+     * @param object a JanitorObject or Java null.
+     * @return a non-null JanitorObject, which might by Janitor.NULL
+     */
+    public static @NotNull JanitorObject nullableObject(@Nullable JanitorObject object) {
+        return object == null ? Janitor.NULL : object;
+    }
+
+    /**
      * Returns a new JDateTime or NULL.
+     *
      * @param dateTime a Java LocalDateTime or Java null
      * @return a new JDateTime or NULL
      */
@@ -341,6 +361,7 @@ public final class Janitor {
 
     /**
      * Returns a new JDateTime.
+     *
      * @param dateTime a Java LocalDateTime
      * @return a new JDateTime
      */
@@ -350,6 +371,7 @@ public final class Janitor {
 
     /**
      * Returns a new JDateTime from the datetime literal, or NULL.
+     *
      * @param text a datetime literal, or Java null
      * @return a new JDateTime from the datetime literal, or NULL
      */
@@ -359,15 +381,17 @@ public final class Janitor {
 
     /**
      * Returns a new JDate from the date literal, or NULL.
+     *
      * @param text a date literal, or Java null
      * @return a new JDate from the date literal, or NULL
      */
-    public static @NotNull JanitorObject nullableDateFromLiteral(@Nullable String text) {
+    public static @NotNull JanitorObject nullableDateFromLiteral(@Nullable String text) throws JsonException {
         return getBuiltins().nullableDateFromLiteral(text);
     }
 
     /**
      * Returns the current date and time as a JDateTime.
+     *
      * @return the current date and time as a JDateTime
      */
     public static @NotNull JDateTime now() {
@@ -376,6 +400,7 @@ public final class Janitor {
 
     /**
      * Returns the current date as a JDate.
+     *
      * @return the current date as a JDate.
      */
     public static @NotNull JDate today() {
@@ -384,6 +409,7 @@ public final class Janitor {
 
     /**
      * Returns a JDate from the LocalDate argument
+     *
      * @param date a LocalDate
      * @return a JDate
      */
@@ -393,6 +419,7 @@ public final class Janitor {
 
     /**
      * Returns a JDate from the LocalDate argument, or NULL when null.
+     *
      * @param date a LocalDate or Java null
      * @return a JDate or NULL
      */
@@ -403,14 +430,23 @@ public final class Janitor {
     /**
      * Returns a JDate from the given year, month and day.
      * The rules for Java's LocalDate::of method apply for the Java implementation!
+     *
      * @param year  the year to represent, from MIN_YEAR to MAX_YEAR
      * @param month the month-of-year to represent, from 1 (January) to 12 (December)
-     * @param day the day-of-month to represent, from 1 to 31
+     * @param day   the day-of-month to represent, from 1 to 31
      * @return that JDate you asked for
      * @throws DateTimeException in case Java's LocalDateTime does not accept the parameters.
      */
     public static @NotNull JDate date(long year, long month, long day) throws DateTimeException {
         return getBuiltins().date(year, month, day);
+    }
+
+    public static @NotNull JanitorObject nullableDateFromJsonString(@Nullable String jsonString) throws JsonException {
+        return getBuiltins().nullableDateFromJsonString(jsonString);
+    }
+
+    public static @NotNull JanitorObject nullableDateTimeFromJsonString(@Nullable String jsonString) throws JsonException {
+        return getBuiltins().nullableDateTimeFromJsonString(jsonString);
     }
 
     /**
@@ -435,11 +471,6 @@ public final class Janitor {
     public static @NotNull JanitorObject numeric(double v) {
         return getBuiltins().numeric(v);
     }
-    /*
-     * This is mainly useful for parsing JSON, were otherwise any ints would be returned as doubles, which some APIs do not like at all.
-     * (E.g. you get a list of things from an APi as 1.0, 2.0, ... and send those back to the same API and *boom* it doesn't like it.
-     * This should not really be our problem, but turns out the one who cares needs to act because the other side will not do anything.)
-     */
 
     /**
      * Returns a numeric object that best represents the double value.
@@ -451,9 +482,15 @@ public final class Janitor {
     public static @NotNull JanitorObject nullableNumeric(Double v) {
         return getBuiltins().nullableNumeric(v);
     }
+    /*
+     * This is mainly useful for parsing JSON, were otherwise any ints would be returned as doubles, which some APIs do not like at all.
+     * (E.g. you get a list of things from an APi as 1.0, 2.0, ... and send those back to the same API and *boom* it doesn't like it.
+     * This should not really be our problem, but turns out the one who cares needs to act because the other side will not do anything.)
+     */
 
     /**
      * Require a boolean value.
+     *
      * @param value the value to check
      * @return the value, if it's a boolean
      * @throws JanitorGlueException if the value is not a boolean
@@ -467,6 +504,7 @@ public final class Janitor {
 
     /**
      * Require a date value.
+     *
      * @param value the value to check
      * @return the value, if it's a date
      * @throws JanitorGlueException if the value is not a date
@@ -480,6 +518,7 @@ public final class Janitor {
 
     /**
      * Require a value to be string.
+     *
      * @param value the value
      * @return the value as string
      * @throws JanitorGlueException if the value is not a string
@@ -493,6 +532,7 @@ public final class Janitor {
 
     /**
      * Require a datetime value.
+     *
      * @param value the value to check
      * @return the value, if it's a datetime
      * @throws JanitorGlueException if the value is not a datetime
@@ -506,6 +546,7 @@ public final class Janitor {
 
     /**
      * Turns a Java Boolean into a Janitor JBool, or returns NULL for Java null.
+     *
      * @param value a Boolean or null
      * @return a JBool or NULL
      */
@@ -515,6 +556,7 @@ public final class Janitor {
 
     /**
      * Gets the user defined environment provider
+     *
      * @return the user defined environment provider
      */
     public static @Nullable JanitorEnvironmentProvider getUserProvider() {
@@ -524,6 +566,7 @@ public final class Janitor {
     /**
      * Sets the user defined environment provider.
      * This method should only ever be used by application code! Library code should <b>never</b> call this.
+     *
      * @param userProvider the user defined environment provider
      */
     public static void setUserProvider(final @NotNull JanitorEnvironmentProvider userProvider) {
@@ -532,6 +575,7 @@ public final class Janitor {
 
     /**
      * Gets the selected automatic environment provider
+     *
      * @return the selected automatic environment provider
      */
     public static @NotNull JanitorEnvironmentProvider getAutomaticProvider() {
@@ -540,6 +584,7 @@ public final class Janitor {
 
     /**
      * Map a java boolean to a JBool.
+     *
      * @param value the value
      * @return the JBool
      */
@@ -549,6 +594,7 @@ public final class Janitor {
 
     /**
      * Retrieves the best applicable enviornment from either the userProvider or the highest priority automatic provider.
+     *
      * @return an environment
      * @throws IllegalStateException when neither a user env nor an automatically provided env are available
      */
@@ -571,6 +617,7 @@ public final class Janitor {
 
     /**
      * Map a nullable java Boolean to a JBool. Null yields FALSE.
+     *
      * @param value the value
      * @return the JBool
      */
@@ -579,9 +626,21 @@ public final class Janitor {
     }
 
     /**
-     * Private constructor, to keep you from creating instances of this singleton / "namespace class".
+     * Compiles a script (source code) to a runnable script (an executable AST in the default implementation).
+     *
+     * @param runtime    a runtime to use [TODO: this is actually only used to retrieve an environment, so this could and should be dropped soon!]
+     * @param moduleName the name of the script module; this will show up in stack traces, so it's a good idea to pick an informative name here if you run lots of scripts
+     * @param source     the actual script source code
+     * @return the script in executable form
+     * @throws JanitorCompilerException on compiler errors
      */
-    private Janitor() {
+    public RunnableScript compile(JanitorRuntime runtime, String moduleName, @Language("Janitor") String source) throws JanitorCompilerException {
+        return runtime.compile(moduleName, source);
+    }
+
+    // TODO: properly document this, and when you're at it make this more useful.
+    public RunnableScript checkCompile(JanitorRuntime runtime, String moduleName, @Language("Janitor") String source) throws JanitorCompilerException {
+        return runtime.checkCompile(moduleName, source);
     }
 
     /**
@@ -591,26 +650,11 @@ public final class Janitor {
     public static final class MetaData {
 
         /**
-         * The class name for an object.
-         */
-        public static MetaDataKey<String> CLASS = new MetaDataKey<>("class", String.class);
-
-        /**
-         * Gives Java code an optional hint what the type of a property is supposed to be.
-         * Turns out that knowing the type of an object is very convenient.
-         * These hints are automatically populated by most Dispatcher methods.
-         */
-        public enum TypeHint {
-            NUMBER, INTEGER, FLOAT, STRING, BOOLEAN, METHOD, DATE, DATETIME, LIST
-        }
-
-        /**
          * This optional meta-data annotation can be used to tell host code (= Java) what a property's type is supposed to be.
          * This can be useful, for example, when host code wants to write an object into a database.
          * Note that these are <b>Janitor</b> types, not Java types, so e.g. a FLOAT could be a Double in Java or a Float...
          */
         public static final MetaDataKey<TypeHint> TYPE_HINT = new MetaDataKey<>("type_hint", TypeHint.class);
-
         /**
          * Meta-Data: HELP for an object or property.
          * <p>
@@ -619,8 +663,6 @@ public final class Janitor {
          * </p>
          */
         public static final MetaDataKey<String> HELP = new MetaDataKey<>("help", String.class);
-
-
         /**
          * Meta-Data: NAME for an object or property.
          * <p>
@@ -629,8 +671,6 @@ public final class Janitor {
          * </p>
          */
         public static final MetaDataKey<String> NAME = new MetaDataKey<>("name", String.class);
-
-
         /**
          * Meta-Data: mark a property as required.
          * <p>
@@ -640,51 +680,56 @@ public final class Janitor {
          * </p>
          */
         public static final MetaDataKey<Boolean> REQUIRED = new MetaDataKey<>("required", Boolean.class);
-
+        /**
+         * The class name for an object.
+         */
+        public static MetaDataKey<String> CLASS = new MetaDataKey<>("class", String.class);
         /**
          * Helper for emitting TS defs for Janutor objects: A property points to a class of this name.
          * To be used where type hints are not possible, because they refer to builtin types only.
          */
         public static MetaDataKey<String> REF = new MetaDataKey<>("ref", String.class);
-
         /**
          * Marks fields that are nullable on the Java side, without specifying anything about script-side nullability.
          */
         public static MetaDataKey<Boolean> HOST_NULLABLE = new MetaDataKey<>("host_nullable", Boolean.class);
-
-
         /**
          * The column name for an object property, in case you're working with a database.
          */
         public static MetaDataKey<String> COLUMN_NAME = new MetaDataKey<>("column_name", String.class);
-
         /**
          * The database table name for an object.
          */
         public static MetaDataKey<String> TABLE_NAME = new MetaDataKey<>("table_name", String.class);
-
         /**
          * The column name of an ID field for an object, e.g. "person_id".
          */
         public static MetaDataKey<String> ID_FIELD = new MetaDataKey<>("id_field", String.class);
-
         /**
          * The column name of a KEY field for an object, e.g. "person_key".
          */
         public static MetaDataKey<String> KEY_FIELD = new MetaDataKey<>("key_field", String.class);
-
-
         /**
          * The maximum length for an object property.
          */
         public static MetaDataKey<Integer> MAX_LENGTH = new MetaDataKey<>("max_length", Integer.class);
 
 
-
         /**
          * Private constructor, to keep you from creating instances of this singleton / "namespace class".
          */
-        private MetaData() {}
+        private MetaData() {
+        }
+
+
+        /**
+         * Gives Java code an optional hint what the type of a property is supposed to be.
+         * Turns out that knowing the type of an object is very convenient.
+         * These hints are automatically populated by most Dispatcher methods.
+         */
+        public enum TypeHint {
+            NUMBER, INTEGER, FLOAT, STRING, BOOLEAN, METHOD, DATE, DATETIME, LIST
+        }
 
     }
 
@@ -719,6 +764,7 @@ public final class Janitor {
         public static @NotNull JBool areNotEquals(final JanitorObject leftValue, final JanitorObject rightValue) {
             return areEquals(leftValue, rightValue).opposite();
         }
+
 
     }